refactor data.py, move pymatgen data stuff in there
