+++ matminer/descriptors/composition.py
@@ -1,16 +1,12 @@
 from __future__ import division, unicode_literals, print_function
 
 import itertools
-import re
-from collections import namedtuple
 
 import numpy as np
+
 from matminer.descriptors.base import AbstractFeaturizer
 from matminer.descriptors.data import magpie_data
 from matminer.descriptors.utils import get_holder_mean
-from pymatgen import Element, Composition
-from pymatgen.core.periodic_table import get_el_sp
-from pymatgen.core.units import Unit
 
 __author__ = 'Jimin Chen, Logan Ward, Saurabh Bajaj, Anubhav jain, Kiran Mathew'
 
@@ -64,39 +60,39 @@ class ElementalAttribute(AbstractFeaturizer):
         mode of descriptors
 
     Args:
-        attributes (list of strings): List of elemental properties to use
+        properties (list of strings): List of elemental properties to use
     """
 
-    def __init__(self, attributes=None):
-        if attributes is None:
-            self.attributes = ["Number", "MendeleevNumber", "AtomicWeight", "MeltingT", "Column",
+    def __init__(self, properties=None):
+        if properties is None:
+            self.properties = ["Number", "MendeleevNumber", "AtomicWeight", "MeltingT", "Column",
                                "Row", "CovalentRadius", "Electronegativity",
                                "NsValence", "NpValence", "NdValence", "NfValence", "NValance",
                                "NsUnfilled", "NpUnfilled", "NdUnfilled", "NfUnfilled", "NUnfilled",
                                "GSvolume_pa", "GSbandgap", "GSmagmom", "SpaceGroupNumber"]
         else:
-            self.attributes = attributes
+            self.properties = properties
 
     def featurize(self, comp):
-        all_attributes = []
+        all_properties = []
 
-        for attr in self.attributes:
-            elem_data = magpie_data.get_data(comp, attr)
+        for attr in self.properties:
+            elem_data = magpie_data.get_property(comp, attr)
 
-            all_attributes.append(min(elem_data))
-            all_attributes.append(max(elem_data))
-            all_attributes.append(max(elem_data) - min(elem_data))
+            all_properties.append(min(elem_data))
+            all_properties.append(max(elem_data))
+            all_properties.append(max(elem_data) - min(elem_data))
 
             prop_mean = sum(elem_data) / len(elem_data)
-            all_attributes.append(prop_mean)
-            all_attributes.append(sum(np.abs(np.subtract(elem_data, prop_mean))) / len(elem_data))
-            all_attributes.append(max(set(elem_data), key=elem_data.count))
+            all_properties.append(prop_mean)
+            all_properties.append(sum(np.abs(np.subtract(elem_data, prop_mean))) / len(elem_data))
+            all_properties.append(max(set(elem_data), key=elem_data.count))
 
-        return all_attributes
+        return all_properties
 
     def generate_labels(self):
         labels = []
-        for attr in self.attributes:
+        for attr in self.properties:
             labels.append("Min %s" % attr)
             labels.append("Max %s" % attr)
             labels.append("Range %s" % attr)
@@ -118,11 +114,11 @@ class ValenceOrbitalAttribute(AbstractFeaturizer):
     def featurize(self, comp):
         num_atoms = comp.num_atoms
 
-        avg_total_valence = sum(magpie_data.get_data(comp, "NValance")) / num_atoms
-        avg_s = sum(magpie_data.get_data(comp, "NsValence")) / num_atoms
-        avg_p = sum(magpie_data.get_data(comp, "NpValence")) / num_atoms
-        avg_d = sum(magpie_data.get_data(comp, "NdValence")) / num_atoms
-        avg_f = sum(magpie_data.get_data(comp, "NfValence")) / num_atoms
+        avg_total_valence = sum(magpie_data.get_property(comp, "NValance")) / num_atoms
+        avg_s = sum(magpie_data.get_property(comp, "NsValence")) / num_atoms
+        avg_p = sum(magpie_data.get_property(comp, "NpValence")) / num_atoms
+        avg_d = sum(magpie_data.get_property(comp, "NdValence")) / num_atoms
+        avg_f = sum(magpie_data.get_property(comp, "NfValence")) / num_atoms
 
         Fs = avg_s / avg_total_valence
         Fp = avg_p / avg_total_valence
@@ -163,8 +159,8 @@ class IonicAttribute(AbstractFeaturizer):
             avg_ionic_char = 0
         else:
             # Get magpie data for each element
-            all_ox_states = magpie_data.get_data(comp, "OxidationStates")
-            all_elec = magpie_data.get_data(comp, "Electronegativity")
+            all_ox_states = magpie_data.get_property(comp, "OxidationStates")
+            all_elec = magpie_data.get_property(comp, "Electronegativity")
             ox_states = []
             elec = []
 
@@ -203,123 +199,6 @@ class IonicAttribute(AbstractFeaturizer):
         return labels
 
 
-def get_pymatgen_descriptor(composition, property_name):
-    """
-    Get descriptor data for elements in a compound from pymatgen.
-
-    Args:
-        composition (str/Composition): Either pymatgen Composition object or string formula,
-            eg: "NaCl", "Na+1Cl-1", "Fe2+3O3-2" or "Fe2 +3 O3 -2"
-            Notes:
-                 - For 'ionic_radii' property, the Composition object must be made of oxidation
-                    state decorated Specie objects not the plain Element objects.
-                    eg.  fe2o3 = Composition({Specie("Fe", 3): 2, Specie("O", -2): 3})
-                 - For string formula, the oxidation state sign(+ or -) must be specified explicitly.
-                    eg.  "Fe2+3O3-2"
-
-        property_name (str): pymatgen element attribute name, as defined in the Element class at
-            http://pymatgen.org/_modules/pymatgen/core/periodic_table.html
-
-    Returns:
-        (list) of values containing descriptor floats for each atom in the compound(sorted by the
-            electronegativity of the contituent atoms)
-
-    """
-    eldata = []
-    # what are these named tuples for? not used or returned! -KM
-    eldata_tup_lst = []
-    eldata_tup = namedtuple('eldata_tup', 'element propname propvalue propunit amt')
-
-    oxidation_states = {}
-    if isinstance(composition, Composition):
-        # check whether the composition is composed of oxidation state decorates species (not just plain Elements)
-        if hasattr(composition.elements[0], "oxi_state"):
-            oxidation_states = dict([(str(sp.element), sp.oxi_state) for sp in composition.elements])
-        el_amt_dict = composition.get_el_amt_dict()
-    # string
-    else:
-        comp, oxidation_states = get_composition_oxidation_state(composition)
-        el_amt_dict = comp.get_el_amt_dict()
-
-    symbols = sorted(el_amt_dict.keys(), key=lambda sym: get_el_sp(sym).X)
-
-    for el_sym in symbols:
-
-        element = Element(el_sym)
-        property_value = None
-        property_units = None
-
-        try:
-            p = getattr(element, property_name)
-        except AttributeError:
-            print("{} attribute missing".format(property_name))
-            raise
-
-        if p is not None:
-            if property_name in ['ionic_radii']:
-                if oxidation_states:
-                    property_value = element.ionic_radii[oxidation_states[el_sym]]
-                    property_units = Unit("ang")
-                else:
-                    raise ValueError("oxidation state not given for {}; It does not yield a unique "
-                                     "number per Element".format(property_name))
-            else:
-                property_value = float(p)
-
-            # units are None for these pymatgen descriptors
-            # todo: there seem to be a lot more unitless descriptors which are not listed here... -Alex D
-            if property_name not in ['X', 'Z', 'group', 'row', 'number', 'mendeleev_no', 'ionic_radii']:
-                property_units = p.unit
-
-        # Make a named tuple out of all the available information
-        eldata_tup_lst.append(eldata_tup(element=el_sym, propname=property_name, propvalue=property_value,
-                                         propunit=property_units, amt=el_amt_dict[el_sym]))
-
-        # Add descriptor values, one for each atom in the compound
-        for i in range(int(el_amt_dict[el_sym])):
-            eldata.append(property_value)
-
-    return eldata
-
-
-def get_composition_oxidation_state(formula):
-    """
-    Returns the composition and oxidation states from the given formula.
-    Formula examples: "NaCl", "Na+1Cl-1",   "Fe2+3O3-2" or "Fe2 +3 O3 -2"
-
-    Args:
-        formula (str):
-
-    Returns:
-        pymatgen.core.composition.Composition, dict of oxidation states as strings
-
-    """
-    oxidation_states_dict = {}
-    non_alphabets = re.split('[a-z]+', formula, flags=re.IGNORECASE)
-    if not non_alphabets:
-        return Composition(formula), oxidation_states_dict
-    oxidation_states = []
-    for na in non_alphabets:
-        s = na.strip()
-        if s != "" and ("+" in s or "-" in s):
-            digits = re.split('[+-]+', s)
-            sign_tmp = re.split('\d+', s)
-            sign = [x.strip() for x in sign_tmp if x.strip() != ""]
-            oxidation_states.append("{}{}".format(sign[-1], digits[-1].strip()))
-    if not oxidation_states:
-        return Composition(formula), oxidation_states_dict
-    formula_plain = []
-    before, after = tuple(formula.split(oxidation_states[0], 1))
-    formula_plain.append(before)
-    for oxs in oxidation_states[1:]:
-        before, after = tuple(after.split(oxs, 1))
-        formula_plain.append(before)
-    for i, g in enumerate(formula_plain):
-        el = re.split("\d", g.strip())[0]
-        oxidation_states_dict[str(Element(el))] = int(oxidation_states[i])
-    return Composition("".join(formula_plain)), oxidation_states_dict
-
-
 if __name__ == '__main__':
 
     import pandas as pd
@@ -329,8 +208,8 @@ if __name__ == '__main__':
 
     for desc in descriptors:
         print(get_pymatgen_descriptor('LiFePO4', desc))
-    print(magpie_data.get_data('LiFePO4', 'AtomicVolume'))
-    print(magpie_data.get_data('LiFePO4', 'Density'))
+    print(magpie_data.get_property('LiFePO4', 'AtomicVolume'))
+    print(magpie_data.get_property('LiFePO4', 'Density'))
     print(get_holder_mean([1, 2, 3, 4], 0))
 
     training_set = pd.DataFrame({"composition": ["Fe2O3"]})
