+++ matminer/descriptors/tests/test_composition_features.py
@@ -3,23 +3,22 @@ from __future__ import unicode_literals, division, print_function
 import unittest
 
 import pandas as pd
-from matminer.descriptors.composition import get_composition_oxidation_state, \
-    get_pymatgen_descriptor, StoichiometricAttribute, ElementalAttribute, ValenceOrbitalAttribute, \
+from matminer.descriptors.composition import StoichiometricAttribute, ElementalAttribute, ValenceOrbitalAttribute, \
     IonicAttribute
-from matminer.descriptors.data import magpie_data
+from matminer.descriptors.data import magpie_data, PymatgenData
 from pymatgen import Composition, Specie
 from pymatgen.util.testing import PymatgenTest
 
 
 class MagpiDataTest(unittest.TestCase):
     def test_string_composition(self):
-        oxs = magpie_data.get_data("LiFePO4", "OxidationStates")
+        oxs = magpie_data.get_property("LiFePO4", "OxidationStates")
         self.assertEqual(oxs, [[1.0],
                                [2.0, 3.0],
                                [-3.0, 3.0, 5.0],
                                [-2.0], [-2.0], [-2.0], [-2.0]])
         lifepo4 = Composition("LiFePO4")
-        oxs_1 = magpie_data.get_data(lifepo4, "OxidationStates")
+        oxs_1 = magpie_data.get_property(lifepo4, "OxidationStates")
         self.assertEqual(oxs, oxs_1)
 
 
@@ -62,33 +61,34 @@ class PymatgenDescriptorTest(unittest.TestCase):
         self.fe2o3_formula_1 = "Fe2+3O3-2"
         self.fe2o3_formula_2 = "Fe2 +3 O3 -2"
         self.lifepo4 = "LiFePO4"
+        self.pmg_data = PymatgenData()
 
     def test_comp_oxstate_from_formula(self):
-        fe2o3_comp_1, fe2o3_oxstates_1 = get_composition_oxidation_state(self.fe2o3_formula_1)
+        fe2o3_comp_1, fe2o3_oxstates_1 = self.pmg_data.get_composition_oxidation_state(self.fe2o3_formula_1)
         oxstates_ans = {'Fe': 3, 'O': -2}
         comp_ans = Composition("Fe2O3")
         self.assertEqual(fe2o3_comp_1, comp_ans)
         self.assertDictEqual(fe2o3_oxstates_1, oxstates_ans)
-        fe2o3_comp_2, fe2o3_oxstates_2 = get_composition_oxidation_state(self.fe2o3_formula_2)
+        fe2o3_comp_2, fe2o3_oxstates_2 = self.pmg_data.get_composition_oxidation_state(self.fe2o3_formula_2)
         self.assertEqual(fe2o3_comp_1, fe2o3_comp_2)
         self.assertDictEqual(fe2o3_oxstates_1, fe2o3_oxstates_2)
-        lifepo4_comp, lifepo4_oxstates = get_composition_oxidation_state(self.lifepo4)
+        lifepo4_comp, lifepo4_oxstates = self.pmg_data.get_composition_oxidation_state(self.lifepo4)
         self.assertEqual(lifepo4_comp, Composition(self.lifepo4))
         self.assertDictEqual(lifepo4_oxstates, {})
 
     def test_descriptor_ionic_radii(self):
-        ionic_radii = get_pymatgen_descriptor(self.nacl_formula_2, "ionic_radii")
+        ionic_radii = self.pmg_data.get_property(self.nacl_formula_2, "ionic_radii")
         self.assertEqual(ionic_radii, [1.16, 1.67])
         with self.assertRaises(ValueError):
-            get_pymatgen_descriptor(self.nacl_formula_1, "ionic_radii")
-        ionic_radii = get_pymatgen_descriptor(self.fe2o3_formula_1, "ionic_radii")
+            self.pmg_data.get_property(self.nacl_formula_1, "ionic_radii")
+        ionic_radii = self.pmg_data.get_property(self.fe2o3_formula_1, "ionic_radii")
         self.assertEqual(ionic_radii, [0.785, 0.785, 1.26, 1.26, 1.26])
 
     def test_descriptor_ionic_radii_from_composition(self):
         cscl = Composition({Specie("Cs", 1): 1, Specie("Cl", -1): 1})
-        ionic_radii = get_pymatgen_descriptor(cscl, "ionic_radii")
+        ionic_radii = self.pmg_data.get_property(cscl, "ionic_radii")
         self.assertEqual(ionic_radii, [1.81, 1.67])
-        ionic_radii_2 = get_pymatgen_descriptor("Cs+1Cl-1", "ionic_radii")
+        ionic_radii_2 = self.pmg_data.get_property("Cs+1Cl-1", "ionic_radii")
         self.assertEqual(ionic_radii, ionic_radii_2)
